Implement resized image cache. Issue #108.
When rendering book covers - we now are trying to use 150x150 images, if they present. Presence checked based on a image cache. Image cache is updated based on `/job/sync_image_cache` command. At the moment it's never called. We need to implement a GCP function that will resize images in Cloud Storage and call the function to update it.