Improve book preview rendering. Issue #109.
When rendering book previews we often need to render only specific narrations. For example latest narration or narration that were narrated by a specific person. To achieve that I added BookForPreview data class that contains Book and list of Narrations to render. Previously we passed Book and now views need to prepare BookForPreview object (or list of such objects) and pass them.